requirements.txt
torch>=1.6
torchvision
numpy
pandas
tqdm
tensorboardX>=1.14
scipy 
opencv-python
clean-fid
labelme
einops
transformers==4.19.2
kornia==0.6
lpips
yaml
Pillow
omegaconf
ftfy